In unix mode mark/remove carriage return
When the buffer is of `dos` fileformat, ignore the CR (^M) characters;
in unix, mark them as trailing white space and delete when saving.

The regex for highlighting depends on vim's default behavior of hiding
the carriage return characters when in dos mode. It will highlight them
either way, but when we want them they will not be visible. Saves a
bunch of `if dos` checks.

Thanks @jyn514 for the suggestion.

Fix #19